Refactor Modules & Make GuildModule scalable

+++ Models/Guild/Roles/IAmRole.cs
@@ -2,11 +2,11 @@
 {
     public class IAmRole
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public ulong GuildId { get; set; }
         public ulong RoleId { get; set; }
-        public int Price { get; set; }
-        public int LevelRequired { get; set; }
+        public ulong Price { get; set; }
+        public ulong LevelRequired { get; set; }
         public ulong RequiredRoleId { get; set; }
     }
 }
