+++ Models/Guild/Roles/LevelRewards.cs
@@ -2,9 +2,9 @@
 {
     public class LevelRewards
     {
-        public int Id { get; set; }
+        public ulong Id { get; set; }
         public ulong GuildId { get; set; }
         public ulong RoleId { get; set; }
-        public int LevelRequired { get; set; }
+        public ulong LevelRequired { get; set; }
     }
 }
